Handle deno parsing cache
