+++ packages/analysis/src/jsts/rules/helpers/dependency-manifests/parsed-manifests.ts
@@ -15,8 +15,10 @@
  * along with this program; if not, see https://sonarsource.com/license/ssal/
  */
 import type { PackageJson } from 'type-fest';
+import ts from 'typescript';
 import yaml from 'yaml';
 import { type File, stripBOM } from '../files.js';
+import type { DenoManifest } from './resolvers/types.js';
 
 export type PnpmWorkspace = {
   packages?: string[];
@@ -26,10 +28,12 @@ export type PnpmWorkspace = {
 
 const parsedPackageJsonCache = new Map<string, PackageJson | undefined>();
 const parsedPnpmWorkspaceCache = new Map<string, PnpmWorkspace | undefined>();
+const parsedDenoManifestCache = new Map<string, DenoManifest | undefined>();
 
 export function clearParsedManifestCache(): void {
   parsedPackageJsonCache.clear();
   parsedPnpmWorkspaceCache.clear();
+  parsedDenoManifestCache.clear();
 }
 
 export function parsePackageJson(file: File): PackageJson | undefined {
@@ -80,5 +84,31 @@ function parsePnpmWorkspaceContent(file: File): PnpmWorkspace | undefined {
   } catch (error) {
     console.debug(`Error parsing pnpm workspace ${file.path}: ${error}`);
     return undefined;
+  }
+}
+
+export function parseDenoManifest(file: File): DenoManifest | undefined {
+  const cached = parsedDenoManifestCache.get(file.path);
+  if (cached || parsedDenoManifestCache.has(file.path)) {
+    return cached;
+  }
+  const parsed = parseDenoManifestContent(file);
+  parsedDenoManifestCache.set(file.path, parsed);
+  return parsed;
+}
+
+function parseDenoManifestContent(file: File): DenoManifest | undefined {
+  try {
+    // ts.parseConfigFileTextToJson handles JSON with comments and trailing commas
+    const parsed = ts.parseConfigFileTextToJson(file.path, stripBOM(file.content.toString()));
+    if (parsed.error) {
+      const message = ts.flattenDiagnosticMessageText(parsed.error.messageText, '\n');
+      console.debug(`Error parsing deno manifest ${file.path}: ${message}`);
+      return;
+    }
+    return parsed.config as DenoManifest;
+  } catch (error) {
+    console.debug(`Error parsing deno manifest ${file.path}: ${error}`);
+    return;
   }
 }
